added a setter and optimized a bit
Made it such that you can change the playlistsize
Optimized the main class such that instead of loading all the movies all the tiem, when editing or deleting movies it will remove/delete in the view and background(db/local) and not load the whole thing over again.
