+++ src/GUI/CONTROLLER/MainViewController.java
@@ -1,16 +1,12 @@
 package GUI.CONTROLLER;
 
-import BE.InputAlert;
-import BE.MusicPlayer;
-import BE.Category;
-import BE.Movie;
+import BE.*;
 import BLL.CategoryManager;
 import BLL.MovieManager;
 import GUI.Main;
 import javafx.beans.property.SimpleStringProperty;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
-import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.fxml.Initializable;
@@ -18,7 +14,6 @@ import javafx.scene.Scene;
 import javafx.scene.control.*;
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
-import javafx.scene.input.MouseEvent;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.BorderPane;
 import javafx.stage.FileChooser;
@@ -78,6 +73,7 @@ public class MainViewController implements Initializable {
     private static final CategoryManager CATEGORY_MANAGER = new CategoryManager();
     private static final MovieManager MOVIE_MANAGER = new MovieManager();
     private static final MusicPlayer musicPlayer = new MusicPlayer();
+    private static final ConvenientUtils CONVENIENT_UTILS = new ConvenientUtils();
 
     /**
      * Constructor
@@ -118,21 +114,7 @@ public class MainViewController implements Initializable {
         selectedCategory();
         setMainViewSize();
         moveMainView();
-        showLowerPane();
-    }
-
-    private void hideLowerPane() {
-        lowerPane.setVisible(false);
-        lowerPane.setMaxHeight(0);
-        lowerPane.setMinHeight(0);
-        lowerPane.setPrefHeight(0);
-    }
-
-    private void showLowerPane(){
-        lowerPane.setVisible(true);
-        lowerPane.setMaxHeight(215);
-        lowerPane.setMinHeight(215);
-        lowerPane.setPrefHeight(215);
+        hideLowerPane();
     }
 
     /**
@@ -190,7 +172,7 @@ public class MainViewController implements Initializable {
         this.moviesInCategoryTable.getSelectionModel().selectedItemProperty().addListener(((observableValue, oldValue, newValue) -> {
             this.selectedMovieInCategory = (Movie) newValue;
             if (selectedMovieInCategory != null) {
-                showLowerPane();
+                //showLowerPane();
                 this.movieTable.getSelectionModel().clearSelection();
             }
         }));
@@ -203,7 +185,7 @@ public class MainViewController implements Initializable {
         this.movieTable.getSelectionModel().selectedItemProperty().addListener(((observableValue, oldValue, newValue) -> {
             this.selectedMovie = (Movie) newValue;
             if (selectedMovie != null) {
-                showLowerPane();
+                //showLowerPane();
                 this.moviesInCategoryTable.getSelectionModel().clearSelection();
             }
         }));
@@ -315,7 +297,6 @@ public class MainViewController implements Initializable {
         try {
             selectedCategory.setCategoryName(newTitle);
             CATEGORY_MANAGER.updateCategory(selectedCategory);
-            reloadCategoryTable();
         } catch (Exception e) {
             e.printStackTrace();
         }
@@ -356,8 +337,8 @@ public class MainViewController implements Initializable {
         if (result.get() == ButtonType.OK) {
             try {
                 CATEGORY_MANAGER.deleteCategory(selectedCategory);
+                categories.remove(selectedCategory);
                 reloadMoviesInCategory();
-                load();
             } catch (Exception e) {
                 e.printStackTrace();
             }
@@ -392,7 +373,7 @@ public class MainViewController implements Initializable {
      * Moves a movie up on the current category.
      */
     public void moveMovieUpInCategoryButton() {
-        this.categoryMovies = FXCollections.observableArrayList(moveInCategory(moviesInCategoryTable.getItems(), selectedMovieInCategory, -1));
+        this.categoryMovies = FXCollections.observableArrayList(CONVENIENT_UTILS.moveInCategory(moviesInCategoryTable.getItems(), selectedMovieInCategory, -1));
         this.moviesInCategoryTable.setItems(categoryMovies);
     }
 
@@ -400,69 +381,8 @@ public class MainViewController implements Initializable {
      * Moves a movie down on the current category
      */
     public void moveMovieDownInCategoryButton() {
-        this.categoryMovies = FXCollections.observableArrayList(moveInCategory(moviesInCategoryTable.getItems(), selectedMovieInCategory, 1));
+        this.categoryMovies = FXCollections.observableArrayList(CONVENIENT_UTILS.moveInCategory(moviesInCategoryTable.getItems(), selectedMovieInCategory, 1));
         this.moviesInCategoryTable.setItems(categoryMovies);
-    }
-
-    /**
-     * Changes the position on the category
-     *
-     * @param listOfMovies the list of movies you want to change
-     * @param movie        the movie
-     * @param pos          the position
-     * @return A list of movies with the new order
-     */
-    public List<Movie> moveInCategory(List<Movie> listOfMovies, Movie movie, int pos) {
-        LinkedList<Movie> linkedMovies = new LinkedList<>(listOfMovies);
-        int index = linkedMovies.indexOf(movie) + pos;
-        if (linkedMovies.size() == 2) {
-            linkedMovies.addLast(linkedMovies.get(0));
-            linkedMovies.remove(linkedMovies.getFirst());
-        }
-        if (linkedMovies.size() > 2) {
-            if (index < 0) {
-                linkedMovies.removeFirstOccurrence(movie);
-                linkedMovies.addLast(movie);
-            } else if (index >= linkedMovies.size()) {
-                linkedMovies.removeLastOccurrence(movie);
-                linkedMovies.addFirst(movie);
-            }
-            if (index >= 0 && index < linkedMovies.size()) {
-                linkedMovies.remove(movie);
-                linkedMovies.add(index, movie);
-            }
-        }
-        return linkedMovies;
-    }
-
-    /**
-     * Saves the category
-     */
-    public void saveCategory() {
-        if (moviesInCategoryTable.getItems() != null) {
-            for (Movie movie : categoryMovies) {
-                try {
-                    CATEGORY_MANAGER.deleteMovieFromCategory(selectedCategory.getCategoryId(), movie.getId());
-                } catch (Exception e) {
-                    e.printStackTrace();
-                }
-            }
-            for (Movie movie : categoryMovies) {
-                try {
-                    CATEGORY_MANAGER.addMoviesToCategory(selectedCategory.getCategoryId(), movie.getId());
-                } catch (Exception e) {
-                    e.printStackTrace();
-                }
-            }
-        }
-    }
-
-    /**
-     * Shuffles the categories
-     */
-    @FXML
-    private void shuffleCategory() {
-        Collections.shuffle(categoryMovies);
     }
 
     /**
@@ -473,8 +393,8 @@ public class MainViewController implements Initializable {
             try {
                 int index = moviesInCategoryTable.getSelectionModel().getFocusedIndex();
                 CATEGORY_MANAGER.deleteMovieFromCategory(selectedCategory.getCategoryId(), selectedMovieInCategory.getId());
-                reloadMoviesInCategory();
-                reloadCategoryTable();
+                categoryMovies.remove(selectedMovie);
+                selectedCategory.setPlaylistSize(selectedCategory.getCategorySize().getValue()-1);
                 moviesInCategoryTable.getSelectionModel().select(index > 0 ? index - 1 : index);
             } catch (Exception e) {
                 e.printStackTrace();
@@ -576,7 +496,7 @@ public class MainViewController implements Initializable {
         if (result.get() == ButtonType.OK) {
             try {
                 MOVIE_MANAGER.deleteMovie(selectedMovie.getId());
-                reloadMoviesInCategory();
+                movies.remove(selectedMovie);
             } catch (Exception e) {
                 e.printStackTrace();
             }
@@ -670,11 +590,25 @@ public class MainViewController implements Initializable {
         allMoviesShown=!allMoviesShown;
     }
 
+    private void hideLowerPane() {
+        lowerPane.setVisible(false);
+        lowerPane.setMaxHeight(0);
+        lowerPane.setMinHeight(0);
+        lowerPane.setPrefHeight(0);
+    }
+
+    private void showLowerPane(){
+        lowerPane.setVisible(true);
+        lowerPane.setMaxHeight(215);
+        lowerPane.setMinHeight(215);
+        lowerPane.setPrefHeight(215);
+    }
+
     public void viewMovieOrWhatever() {
     }
 
     public void imgChange() {
         movieImg.setImage(new Image(""));
-        System.out.println("WOW THE IMAGE CHANGED");
+        System.out.println("WOW THE IMAGE CHANGED... or is yet to be implemented");
     }
 }
